fix: Challenge 02- created separate components
